requirements.txt
Django
djangorestframework
python-dotenv
bardapi
twilio